# docs/swagger.yaml
basePath: /
definitions:
...
  models.RoomDTO:
    properties:
      capacity:
        type: integer
      location:
        type: string
      name:
        type: string
    type: object
host: localhost:9010
info:
  contact: {}
  description: API documentation for Meeting Room Booking system
  title: Meeting Room Booking API
  version: "1.0"
paths:
...
  /rooms:
    get:
      description: Retrieves all rooms along with their bookings and booked employees
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.RoomDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get list of all rooms
      tags:
      - Rooms
    post:
      consumes:
      - application/json
      description: Adds a new meeting room to the system
      parameters:
      - description: Room details
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/models.RoomDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RoomDTO'
        "400":
          description: Invalid JSON or bad request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new room
      tags:
      - Rooms
  /rooms/{id}:
    put:
      consumes:
      - application/json
      description: Update details of existing room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Room details to update
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/models.RoomDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoomDTO'
        "400":
          description: Invalid JSON or bad request
          schema:
            type: string
        "404":
          description: Room not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update room details
      tags:
      - Rooms
swagger: "2.0"